api/requirements.txt
connexion == 1.1.9
python_dateutil == 2.6.0
setuptools >= 21.0.0
boto3 >= 1.5.1